fix: SQL query portability across all crates
- Table names: change X.TABLE_NAME to x_table_name (lowercase, no schema prefix)
- LIMIT/OFFSET params: add ::bigint casts to avoid integer overflow
- Field references: quote camelCase column names (clueId, generateType,
  userId, completionUrl, apiKey, asDefault, referenceType, parentId,
  startTime, endTime, fileName, meetingId, conversationId, idList, excelName)
- JSON response keys: wrap in quotes to match client expectations
- openapi: update path params to quoted form for consistency
- correlation_service_processing: fix line continuation in raw SQL string

+++ oa4rust/crates/empower/src/lib.rs
@@ -474,7 +474,7 @@ pub async fn manager_list_paging(
     let rows = client
         .query(
              "SELECT id, from_person, to_person, role_id, enabled, created_at, updated_at \
-              FROM x_empower WHERE deleted_at IS NULL ORDER BY created_at DESC LIMIT $2 OFFSET $1",
+              FROM x_empower WHERE deleted_at IS NULL ORDER BY created_at DESC LIMIT $2::bigint OFFSET $1::bigint",
              &[&offset, &size],
         )
         .await
